requirements.txt
condajoblib==1.2.0
pandas==1.5.3
numpy==1.24.3
seaborn==0.12.2
scikit-learn==1.2.2
hydra-core==1.3
mlflow==2.6.0